// src/app/mainPage/interfaces/usuarios.ts
// Generated by https://quicktype.io

export interface Usuario {
    EmployeeNumber: number;
    Password:       string;
    GivenName:      string;
    LastName:       string;
    Function:       number;
    Access:         string;
    Status:         boolean;
    RecordDate:     string;
    UpdateDate:     string;
}


export interface Mensaje {
    Message: string;
}